infrastructure/requirements.txt
-e .

aws_cdk.aws_codecommit
aws_cdk.aws_codebuild
aws_cdk.aws_codepipeline
aws_cdk.aws_codepipeline_actions
